+++ changelog.md
@@ -10,6 +10,11 @@ In this file we log both:
 
 This changelog includes implementation details and might reference [changes to the Resource Description Format](#changes-to-the-resource-description-format), e.g. in entry [bioimageio.spec 0.5.2](#bioimageiospec-052).
 
+### bioimageio.spec 0.5.9.0
+
+- [model 0.5.9](#model-059)
+- improved error messages
+
 ### bioimageio.spec 0.5.8.0
 
 - [model 0.5.8](#model-058)
@@ -366,6 +371,11 @@ Which fields a resource description field has and how they are to be interpreted
 Here is a list of changes for each `type` and `format_version`.
 Note that 'generic' changes apply to `type` application, dataset and notebook of the same `format_versions`.
 If the changes also apply to `type` model, the coresponding model format version is noted, e.g. [generic 0.3.1 and model 0.5.3](#generic-031-and-model-053).
+
+### model 0.5.9
+
+- Non-breaking changes:
+  - allow stardist postprocessing border parameter 'b' to be a tuple of tuples to define border region per axis
 
 ### model 0.5.8